添加了json解析parsetools

+++ COToolsSample/COToolsSample/TestDomain.h
@@ -10,7 +10,7 @@
 
 @interface TestDomain : COBaseDomain<NSCoding>
 @property(nonatomic,strong)NSString  * name;
-@property(nonatomic,assign)int age;
+@property(nonatomic,assign)float age;
 @property(nonatomic,strong)NSString  * address;
 @property(nonatomic,strong)TestDomain  * childDomain;
